requirements.txt
Flask~=1.1.2
requests~=2.25.1
lxml~=4.6.2
beautifulsoup4~=4.9.3
WTForms~=2.3.3
flask_ckeditor
flask_login
Werkzeug~=1.0.1
flask_sqlalchemy
SQLAlchemy~=1.3.23
gravatar~=0.1
flask_gravatar
python-dotenv~=0.15.0